requirements.txt
aiogram==3.0.0
Jinja2==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pyscopg2-binary
